refactor: history 최종 refactoring
