+++ src/History/components/HistoryProgressBar.tsx
@@ -1,4 +1,5 @@
 import styled from '@emotion/styled';
+
 interface ProgressBarProps {
   currentProgress: number;
   maximumProgress?: number;
