+++ src/History/type/historyData.ts
@@ -1,36 +1,39 @@
-export interface IObjective {
+export type filterOptionTypes = '최신순' | '오래된 순' | '달성률 순';
+export type selectedThemeTypes =
+  | '성장'
+  | '건강'
+  | '생산성'
+  | '라이프스타일'
+  | '경제'
+  | '셀프케어'
+  | undefined;
+
+export type ObjectiveTypes = {
   objId: number;
-  title: string;
+  objTitle: string;
   objCategory: string;
-  progress: number;
+  objProgress: number;
   objPeriod: string;
   objIdx?: number;
+  isLast?: boolean;
   krList: IKeyResult[];
-}
-export interface IKeyResult {
+};
+
+type IKeyResult = {
   krId: number;
   krIdx: number;
   krTitle: string;
   krProgress: number;
   taskList: ITask[];
-}
+};
 
-export interface ITask {
+type ITask = {
   taskId: number;
   taskIdx: number;
   taskTitle: string;
-}
+};
 
-export interface Group {
+export type GroupTypes = {
   year: number;
-  objList: IObjective[];
-}
-
-export interface IObjectiveDataProps {
-  groups: Group[];
-  years: {
-    [year: string]: number;
-  };
-  categories: string[];
-  onThemeSelect: (selectedTheme: string) => void;
-}
+  objList: ObjectiveTypes[];
+};
